requirements.txt
numpy==1.21.5
matplotlib==3.5.2
scipy==1.9.1
beautifulsoup4==4.11.1
requests==2.28.1
pandas==1.4.4
tabula-py==2.6.0